+++ src/topology/mod.rs
@@ -6,6 +6,6 @@ mod edge;
 mod shape;
 mod topology;
 
-pub use edge::{DirectedEdge, EdgeExtensions, EdgeKind};
+pub use edge::{DirectedEdge, EdgeKind};
 pub use shape::*;
 pub use topology::{Topology, TopologyMetrics, ValidationLevel};